refactor(core): move ContractAddress function to transaction.go, delete contract.go
