--- core/contract.go
@@ -1,26 +1,0 @@
-package core
-
-import (
-	"github.com/NethermindEth/juno/core/crypto"
-	"github.com/NethermindEth/juno/core/felt"
-)
-
-// ContractAddress computes the address of a Starknet contract.
-func ContractAddress(
-	callerAddress,
-	classHash,
-	salt *felt.Felt,
-	constructorCallData []*felt.Felt,
-) felt.Felt {
-	prefix := felt.FromBytes[felt.Felt]([]byte("STARKNET_CONTRACT_ADDRESS"))
-	callDataHash := crypto.PedersenArray(constructorCallData...)
-
-	// https://docs.starknet.io/architecture-and-concepts/smart-contracts/contract-address/
-	return crypto.PedersenArray(
-		&prefix,
-		callerAddress,
-		salt,
-		classHash,
-		&callDataHash,
-	)
-}
